@@ -1,7 +1,7 @@
 package i2p.bote.android;
 
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.Toolbar;
 
 public class NetworkInfoActivity extends BoteActivityBase {
     @Override
