@@ -3,8 +3,8 @@ package i2p.bote.android.config;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceFragmentCompat;
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
 
 import i2p.bote.android.R;
 
