@@ -1,7 +1,7 @@
 package i2p.bote.android.config.util;
 
 import android.os.Bundle;
-import android.support.v7.preference.EditTextPreferenceDialogFragmentCompat;
+import androidx.preference.EditTextPreferenceDialogFragmentCompat;
 import android.text.InputType;
 import android.view.View;
 import android.widget.EditText;
