@@ -1,7 +1,7 @@
 package i2p.bote.android.config.util;
 
 import android.content.Context;
-import android.support.v7.preference.EditTextPreference;
+import androidx.preference.EditTextPreference;
 import android.util.AttributeSet;
 
 public class SummaryEditTextPreference extends EditTextPreference {
