@@ -1,7 +1,7 @@
 package i2p.bote.android.util;
 
 import android.content.Context;
-import android.support.v4.content.AsyncTaskLoader;
+import androidx.loader.content.AsyncTaskLoader;
 
 public abstract class BetterAsyncTaskLoader<T> extends AsyncTaskLoader<T> {
     protected T mData;
