@@ -18,9 +18,10 @@
 package i2p.bote.android.util;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.view.ActionMode;
-import android.support.v7.widget.RecyclerView;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.view.ActionMode;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Pair;
 import android.util.SparseBooleanArray;
 import android.view.Menu;
@@ -130,7 +131,7 @@ public class MultiSelectionUtil {
 
         /**
          * This method should be called from your {@link AppCompatActivity} or
-         * {@link android.support.v4.app.Fragment Fragment} to allow the controller to restore any
+         * {@link Fragment Fragment} to allow the controller to restore any
          * instance state.
          *
          * @param savedInstanceState - The state passed to your Activity or Fragment.
@@ -151,7 +152,7 @@ public class MultiSelectionUtil {
         /**
          * This method should be called from
          * {@link AppCompatActivity#onSaveInstanceState(android.os.Bundle)} or
-         * {@link android.support.v4.app.Fragment#onSaveInstanceState(android.os.Bundle)
+         * {@link Fragment#onSaveInstanceState(android.os.Bundle)
          * Fragment.onSaveInstanceState(Bundle)} to allow the controller to save its instance
          * state.
          *
