@@ -131,7 +131,7 @@ import i2p.bote.status.StatusListener;
  */
 public class I2PBote implements NetworkStatusSource, EmailFolderManager, MailSender, PasswordVerifier {
     public static final int PROTOCOL_VERSION = 4;
-    private static final String APP_VERSION = "0.4.8";
+    private static final String APP_VERSION = "0.4.8.1";
     private static final int STARTUP_DELAY = 3;   // the number of minutes to wait before connecting to I2P (this gives the router time to get ready)
     private static volatile I2PBote instance;
     
